changed direct_referral type to Long, and added direct_referral id to added new user based on passed in direct_referral username in add new user end point

+++ src/main/java/com/eden/edenbe/User.java
@@ -27,7 +27,7 @@ public class User {
     private Integer points;
     private String package_type;
     private BigDecimal money_amount;
-    private Integer direct_referral;
+    private Long direct_referral;
 
     public User() {
     }
@@ -50,7 +50,7 @@ public class User {
         int points,
         String package_type,
         BigDecimal new_money_amount,
-        Integer direct_referral
+        Long direct_referral
     ) {
         this.id = id;
         this.username = username;
@@ -133,7 +133,7 @@ public class User {
         return this.package_type;
     }
     public BigDecimal getMoney_amount() { return this.money_amount; }
-    public Integer getDirect_referral() { return this.direct_referral; }
+    public Long getDirect_referral() { return this.direct_referral; }
 
     /*
     * Setters:
@@ -181,5 +181,5 @@ public class User {
         this.package_type = newPackageType;
     }
     public void setMoney_amount(BigDecimal new_money_amount) { this.money_amount = new_money_amount.setScale(2, RoundingMode.HALF_UP); }
-    public void setDirect_referral(Integer direct_referral) { this.direct_referral = direct_referral; }
+    public void setDirect_referral(Long direct_referral) { this.direct_referral = direct_referral; }
 }
